requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.9.0
azure-functions==1.23.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==45.0.6
distro==1.9.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gevent==25.5.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
MarkupSafe==3.0.2
numpy==2.3.2
openai==1.99.9
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
regex==2025.7.34
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
whitenoise==6.9.0
zope.event==5.1.1
zope.interface==7.2
